feat: replace QuestionnaireSettings
